discord-bot/pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = "Discord Lotto Reminder Bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.4",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "apscheduler>=3.11.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.3",
]

[project.scripts]
discord-bot = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
